BOCM-486 - adds tests for get endpoint
