+++ apps/document-storage/package.json
@@ -8,7 +8,8 @@
     },
     "type": "module",
     "scripts": {
-        "dev": "nodemon -e js ./src/server.js"
+        "dev": "nodemon -e js ./src/server.js",
+        "test": "cross-env NODE_ENV=test ava"
     },
     "dependencies": {
         "@azure/storage-blob": "^12.1.2",
@@ -20,5 +21,11 @@
         "morgan": "^1.10.0",
         "multer": "^1.4.1",
         "crypto-js": "4.1.1"
+    },
+    "devDependencies": {
+        "ava": "^4.2.0",
+        "cross-env": "^7.0.3",
+        "sinon": "13.0.2",
+        "supertest": "6.2.2"
     }
 }
