+++ apps/document-storage/src/server/app/validator.js
@@ -1,0 +1,4 @@
+import multer from 'multer';
+
+const upload = multer({ storage: multer.memoryStorage() })
+export const validateDocumentUpload = upload.single('file')
